Refactor Note, Book interface
