+++ bluemog-client/src/components/molecules/Rating/index.tsx
@@ -1,27 +1,19 @@
-import React, { ComponentProps } from 'react';
+import React from 'react';
 import { css } from '@emotion/react';
-import Image from 'components/molecules/Image';
-import { Note } from 'components/molecules/NoteCard';
 import styled from '@emotion/styled';
-import ratingImage from '../../../../public/assets/images/rating.svg';
+import { Note } from 'types/note';
+import Star from 'components/molecules/Star';
 
-type ImageProps = ComponentProps<typeof Image>;
 type RatingType = Note['rating'];
 
 export interface Props {
   rating: RatingType;
 }
 
-const StyledImage = styled(Image)`
+const StyledStar = styled(Star)`
   width: 25px;
-  height: 25px;
   margin-right: 5px;
 `;
-
-const starImage: ImageProps = {
-  imageUrl: ratingImage,
-  alt: 'Rating',
-};
 
 function Rating({ rating }: Props) {
   const componentCSS = css`
@@ -30,13 +22,18 @@ function Rating({ rating }: Props) {
     width: 145px;
   `;
 
+  let num: number = rating;
+  if (num > 5) {
+    num = 5;
+  }
+
   return (
     <div css={componentCSS}>
       {[
-        ...Array(rating),
+        ...Array(num),
       ].map((value: undefined, index: number) => (
         // eslint-disable-next-line react/no-array-index-key
-        <StyledImage {...starImage} key={index} />))}
+        <StyledStar key={index} />))}
     </div>
   );
 }
