+++ bluemog-client/src/types/book.ts
@@ -1,0 +1,6 @@
+export interface Book {
+  title: string;
+  writer: string;
+  publishing: string;
+  imageUrl: string;
+}
