+++ bluemog-client/src/types/note.ts
@@ -1,0 +1,7 @@
+export interface Note {
+  id: number;
+  title: string;
+  rating: number;
+  text: string;
+  imageUrl: string;
+}
